data logic update, splash screen update
