+++ app/src/main/java/com/mtdagar/horoscopepredictions/Networking.kt
@@ -5,14 +5,18 @@ import android.view.View
 import com.androidnetworking.AndroidNetworking
 import com.androidnetworking.common.ANResponse
 import com.androidnetworking.common.Priority
+import com.androidnetworking.error.ANError
+import com.androidnetworking.interfaces.JSONObjectRequestListener
 import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+import com.mtdagar.horoscopepredictions.model.Horo
 import com.mtdagar.horoscopepredictions.model.HoroStory
 import omari.hamza.storyview.model.MyStory
 import org.json.JSONObject
 import java.util.*
 
 
-class Networking constructor(var progressBar: View) {
+class Networking constructor() {
 
     interface NetworkingInterface {
         fun onResponse(sign: String, list: ArrayList<MyStory>, horoObject: HoroStory)
@@ -20,7 +24,7 @@ class Networking constructor(var progressBar: View) {
         fun onError(message: String)
     }
 
-    private suspend fun fetchData(sign: String, day: String): HoroStory? {
+    suspend fun fetchData(sign: String, day: String): HoroStory {
         val response: ANResponse<JSONObject> =
             AndroidNetworking.post("https://sameer-kumar-aztro-v1.p.rapidapi.com/?sign=$sign&day=$day")
                 .addHeaders(
@@ -33,13 +37,23 @@ class Networking constructor(var progressBar: View) {
                 .executeForJSONObject() as ANResponse<JSONObject>
 
         if (response.isSuccess) {
-            val result = response.result
+            var result = response.result
+            Log.i("Result", result.toString())
             val gson = Gson()
-            return gson.fromJson(result.toString(), HoroStory::class.java)
-        }else{
-            Log.i("Response failed: ", response.toString())
-            return null
+            response.result.put("day", day)
+            response.result.put("sign", sign)
+
+            val horoData = gson.fromJson(result.toString(), HoroStory::class.java)
+
+            Log.i("NetworkResponse", horoData.toString())
+            return horoData
+
+        } else {
+            Log.i("Response failed: ", response.error.toString())
+            return HoroStory()
         }
+
+
     }
 
     fun getStories(Sign: String, networkingInterface: NetworkingInterface) {
@@ -93,8 +107,8 @@ class Networking constructor(var progressBar: View) {
                         "$day: ${horoData.description}"
                     )
                 )
-            }else{
-             Log.i("Response failed: ", response.toString())
+            } else {
+                Log.i("Response failed: ", response.toString())
             }
         }
